backend/requirements.txt
# FastAPI 核心
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

# 文件处理
python-docx==1.1.0
Pillow==11.0.0

# 搜索引擎
whoosh==2.7.4
jieba==0.42.1

# HTTP客户端
aiohttp==3.9.3
requests==2.31.0

# 算法库
scikit-learn==1.4.0
numpy==1.26.3

# 工具库
python-dateutil==2.8.2
